Create interrupts module
