+++ kernel/src/interrupts/idt.rs
@@ -2,13 +2,13 @@
 // https://wiki.osdev.org/Interrupts_tutorial
 // https://wiki.osdev.org/Exceptions
 
-use crate::user_program::syscall;
 use arbitrary_int::{u2, u4};
 use bitbybit::bitfield;
 use core::{arch::asm, mem::size_of};
 
-use crate::threading::scheduling;
-use crate::timer;
+use crate::interrupts::intr_handler::{
+    page_fault_handler, syscall_handler, timer_interrupt_handler, unhandled_handler,
+};
 
 #[repr(align(8))]
 #[bitfield(u64, default = 0)]
@@ -39,82 +39,6 @@ static mut IDT: [GateDescriptor; IDT_LEN] = [GateDescriptor::DEFAULT; IDT_LEN];
 // horribly dangerous... The current behaviour is currently safe fine for cases
 // where we're entering a handler from usermode though, because when doing that
 // we get the new stack from the TSS.
-
-#[naked]
-unsafe extern "C" fn unhandled_handler() -> ! {
-    fn inner() -> ! {
-        panic!("unhandled interrupt");
-    }
-
-    asm!(
-        "call {}",
-        sym inner,
-        options(noreturn),
-    );
-}
-
-#[naked]
-unsafe extern "C" fn page_fault_handler() -> ! {
-    unsafe fn inner(error_code: u32, return_eip: usize) -> ! {
-        let vaddr: usize;
-        asm!("mov {}, cr2", out(reg) vaddr);
-        panic!("page fault with error code {error_code:#b} occurred when trying to access {vaddr:#X} from instruction at {return_eip:#X}");
-    }
-
-    asm!(
-        "call {}",
-        sym inner,
-        options(noreturn),
-    );
-}
-
-#[naked]
-unsafe extern "C" fn syscall_handler() -> ! {
-    asm!(
-        "
-        // Push arguments to stack.
-        push edx
-        push ecx
-        push ebx
-        push eax
-
-        // TODO: We need to define what our syscall ABI is allowed to clobber
-        // and what it must preserve, then actually do that. We should also
-        // investigate what actual OSs do to ensure that we're not leaking
-        // sensitive kernel data, even if we are respecting our ABI.
-
-        call {}
-        // eax will contain the handler's return value, which is where it should
-        // remain when we return to the program.
-
-        add esp, 16 // Drop arguments from stack.
-
-        iretd
-        ",
-        sym syscall::handler,
-        options(noreturn),
-    );
-}
-
-#[naked]
-unsafe extern "C" fn timer_interrupt_handler() -> ! {
-    asm!(
-        "
-        // Push IRQ0 value onto the stack.
-        push 0x0
-        call {} // Update system clock
-        call {} // Send EOI signal to PICs
-        call {} // Yield process
-
-        add esp, 4 // Drop arguments from stack
-        iretd
-        ",
-        sym timer::step_sys_clock,
-        sym timer::send_eoi,
-        sym scheduling::scheduler_yield_and_continue,
-        options(noreturn),
-    );
-}
 
 static mut IDT_DESCRIPTOR: IDTDescriptor = IDTDescriptor {
     size: size_of::<[GateDescriptor; IDT_LEN]>() as u16 - 1,
