+++ kernel/src/interrupts/mod.rs
@@ -1,3 +1,9 @@
+pub mod idt;
+pub mod timer;
+
+mod intr_handler;
+pub(crate) mod mutex_irq;
+
 use core::{
     arch::asm,
     sync::atomic::{compiler_fence, Ordering},
@@ -15,10 +21,10 @@ pub fn intr_get_level() -> IntrLevel {
     let flags: u32;
     unsafe {
         asm!(
-            "pushfd",
-            "mov {}, [esp]",
-            "popfd",
-            out(reg) flags
+        "pushfd",
+        "mov {}, [esp]",
+        "popfd",
+        out(reg) flags
         );
     }
 
@@ -26,29 +32,6 @@ pub fn intr_get_level() -> IntrLevel {
         IntrLevel::IntrOn
     } else {
         IntrLevel::IntrOff
-    }
-}
-
-/// A guard for withholding interrupts.
-#[derive(Default)]
-pub struct InterruptsGuard(bool);
-
-impl !Send for InterruptsGuard {}
-
-/// Prevents interrupts from occuring until the the `InterruptsGuard` is dropped.
-/// After it is dropped, the interrupts are returned to the previous state.
-pub fn hold_interrupts() -> InterruptsGuard {
-    let enabled = intr_get_level() == IntrLevel::IntrOn;
-    let retval = InterruptsGuard(enabled);
-    intr_disable();
-    retval
-}
-
-impl Drop for InterruptsGuard {
-    fn drop(&mut self) {
-        if self.0 {
-            intr_enable();
-        }
     }
 }
