+++ kernel/src/sync/mod.rs
@@ -1,5 +1,2 @@
-pub mod intr;
-#[allow(unused)]
-pub mod irq;
 #[allow(unused)]
 pub mod mutex;
